requirements.txt
numpy
pandas
seaborn
opencv-python
matplotlib
ipyfilechooser
scipy
tensorflow
torch
keras
scikit-learn
mlflow
Pillow
mlflow
joblib
huggingface
transformers